+++ NEWS.md
@@ -1,5 +1,7 @@
 # raadfiles dev
 
+* New REMA functions, and data set for the polygons. WIP
+
 * Slight speedup for package load by using text file and vroom and read/write the file database. 
 
 * Move to REMA 1.1, 8m, 100m, 200m.